b200: Fix compiler warnings related to type conversions
These compiler warnings mostly pop up on MSVC. Most of them are due to
inconsistent usage of size_t, uint{8,16,32}_t, and even int.
This commit changes types mostly such that variables have the correct
type to begin with, although it also contains a few explicit type-casts.

+++ host/lib/usrp/b200/b200_impl.cpp
@@ -1262,9 +1262,9 @@ void b200_impl::update_atrs(void)
         const bool enb_rx = bool(perif.rx_streamer.lock());
         const bool enb_tx = bool(perif.tx_streamer.lock());
         const bool is_rx2 = perif.ant_rx2;
-        const size_t rxonly = (enb_rx)? ((is_rx2)? STATE_RX1_RX2 : STATE_RX1_TXRX) : STATE_OFF;
-        const size_t txonly = (enb_tx)? (STATE_TX1_TXRX) : STATE_OFF;
-        size_t fd = STATE_OFF;
+        const uint32_t rxonly = (enb_rx)? ((is_rx2)? STATE_RX1_RX2 : STATE_RX1_TXRX) : STATE_OFF;
+        const uint32_t txonly = (enb_tx)? (STATE_TX1_TXRX) : STATE_OFF;
+        uint32_t fd = STATE_OFF;
         if (enb_rx and enb_tx) fd = STATE_FDX1_TXRX;
         if (enb_rx and not enb_tx) fd = rxonly;
         if (not enb_rx and enb_tx) fd = txonly;
@@ -1280,9 +1280,9 @@ void b200_impl::update_atrs(void)
         const bool enb_rx = bool(perif.rx_streamer.lock());
         const bool enb_tx = bool(perif.tx_streamer.lock());
         const bool is_rx2 = perif.ant_rx2;
-        const size_t rxonly = (enb_rx)? ((is_rx2)? STATE_RX2_RX2 : STATE_RX2_TXRX) : STATE_OFF;
-        const size_t txonly = (enb_tx)? (STATE_TX2_TXRX) : STATE_OFF;
-        size_t fd = STATE_OFF;
+        const uint32_t rxonly = (enb_rx)? ((is_rx2)? STATE_RX2_RX2 : STATE_RX2_TXRX) : STATE_OFF;
+        const uint32_t txonly = (enb_tx)? (STATE_TX2_TXRX) : STATE_OFF;
+        uint32_t fd = STATE_OFF;
         if (enb_rx and enb_tx) fd = STATE_FDX2_TXRX;
         if (enb_rx and not enb_tx) fd = rxonly;
         if (not enb_rx and enb_tx) fd = txonly;
